refactor(rack): extract RMS simulator
